grobid/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
grobid-client-python==0.0.9
idna==3.10
lxml==5.3.1
requests==2.32.3
urllib3==2.3.0
